+++ crates/cxx-qt-gen/test_inputs/shebang.rs
@@ -1,6 +1,6 @@
 #!/usr/bin/
 // Test for qtitem and qtfile, for code coverage
-#[cxx_qt::bridge]
+#[cxx_qt::bridge(cxx_file_stem = "shebang")]
 mod ffi {
     extern "RustQt" {
         #[qobject]